% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coauthorship.R
\docType{data}
\name{coauthorship}
\alias{coauthorship}
\title{Coauthorship network}
\format{Object of class igraph of size 10114, undirected.}
\source{
Own calculation
}
\description{
Coauthorship network of scientists from University of Warsaw and their
externa co-workers. Relation indicates pairs of people who has authored
at least one publication together.
}
\details{
Node attributes include authors' direct affiliation and also their affiliation
to faculty.
}
