% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{earnings}
\alias{earnings}
\title{Subset of Polish General Social Survey data with data on incomes}
\format{SPSS system file}
\source{
Polish Social Science Data Archive (ADS)
}
\description{
Subset of Polish General Social Survey data with data on incomes
}
\references{
PGSS reference
}
\keyword{datasets}

