% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Input_output.R
\docType{data}
\name{Input_output}
\alias{Input_output}
\title{Input-output table for U.S. economy}
\format{22-by-22 numeric matrix with dimnames.}
\source{
TODO
}
\description{
Based on data from TODO
}
\references{
TODO: Bojanowski () "Industrial structure and interfirm collaboration".
}
