% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{isnar}
\alias{isnar}
\alias{isnar-package}
\title{Introduction to Social Network Analysis with R}
\description{
Functions and data accompanying the workshop "Introduction to Social Network
Analysis with R" at annual INSNA Sunbelt conferences.
}
\details{
Among other things, it includes a lot of things.
}

