% Generated by roxygen2 (4.0.1): do not edit by hand
\name{orwg}
\alias{orwg}
\alias{orwg.default}
\alias{orwg.igraph}
\alias{orwg.table}
\title{Odds ratio of existence of within-group ties}
\usage{
orwg(object, ...)

\method{orwg}{table}(object, ...)

\method{orwg}{igraph}(object, vattr, ...)

\method{orwg}{default}(object, ...)
}
\arguments{
\item{object}{R object, see Details for available methods}

\item{...}{other arguments passed to/from other methods}

\item{vattr}{character scalar or any vector, name of the vertex attribute or
the attribute itself (as a vector)}
}
\value{
Numeric value of the measure.
}
\description{
Odds ratio for connected, as opposed to disconnected, dyads depending
whether it is between- or within-group, i.e. how much more likely the dyad
will be connected if it is within-group.
}
\details{
The measure takes values, like all odds ratios, from (0; Inf).

Method for mixing matrices.

Method for igraphs
}
\examples{
orwg(Wnet, "gender")
}
\references{
Moody, Jim (2001) "Race, school integration, and friendship segregation in
America", American Journal of Sociology, 107(3):679--377
}
\seealso{
Other segregation.measures: \code{\link{assort}},
  \code{\link{assort.default}},
  \code{\link{assort.igraph}}, \code{\link{assort.table}};
  \code{\link{coleman}}, \code{\link{coleman.default}},
  \code{\link{coleman.igraph}},
  \code{\link{coleman.table}}; \code{\link{ei}},
  \code{\link{ei.default}}, \code{\link{ei.igraph}},
  \code{\link{ei.table}}; \code{\link{freeman}},
  \code{\link{freeman.default}},
  \code{\link{freeman.igraph}},
  \code{\link{freeman.table}}; \code{\link{gamix}},
  \code{\link{gamix.default}}, \code{\link{gamix.igraph}},
  \code{\link{gamix.table}}; \code{\link{smi}},
  \code{\link{smi.default}}, \code{\link{smi.igraph}},
  \code{\link{smi.table}}; \code{\link{ssi}}
}

