% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_redundancy.R
\name{pairwise_redundancy}
\alias{pairwise_redundancy}
\alias{pr_invdistance}
\alias{pr_sharedp}
\title{Pairwise tie redundancy}
\usage{
pairwise_redundancy(g, v = igraph::V(g))

pr_invdistance(g, v = igraph::V(g))

pr_sharedp(g, v = igraph::V(g))
}
\arguments{
\item{g}{graph}

\item{v}{nodes to be treated as egos}
}
\value{
Every function returns a data frame with columns \code{ego}, \code{v1},
\code{v2} with codes for, respectively, ego, alter1, and alter2. Additional column
depends on function and it is:
\describe{
\item{invdis}{inverse distance returned by \code{pr_invdistance}}
\item{sharedp}{number of shared partners returned by \code{pr_sharedp}}
\item{pr_redundancy}{the two above combined returned by \code{pairwise_redundancy}}
}
}
\description{
Measuring relative redundancy of ties.
}
\details{
\code{pairwise_redundancy} computes a combination of the two measures:
}
