% Generated by roxygen2 (4.0.1): do not edit by hand
\name{coleman}
\alias{coleman}
\alias{coleman.array}
\alias{coleman.igraph}
\title{Coleman's homophily index}
\usage{
coleman(object, ...)

\method{coleman}{array}(object, gsizes = NULL, loops = FALSE, ...)

\method{coleman}{igraph}(object, vattr, ...)
}
\arguments{
\item{object}{R object, see available methods}

\item{...}{other arguments}

\item{gsizes}{numeric vector of group sizes}

\item{loops}{logical, whether loops are allowed}

\item{vattr}{character, vertex attribute}
}
\value{
Vector of numeric values of the index for each group
}
\description{
Colemans's homphily index for directed networks.
}
\details{
Coleman's homophily index computes homophily scores for each group
defined by a vertex attribute.

\code{object} can be a mixing matrix:
\enumerate{
\item{either only the contact layer (square matrix), which requires
supplying the \code{gsizes} argument with the sizes of the groups}
\item{full mixing matrix as a three dimensional array}
}

\code{object} can be of class "igraph"
}
\examples{
coleman(as.directed(Wnet, "mutual"), "gender")

coleman(as.directed(EFnet, "mutual"), "type")
}
\references{
Coleman, J. (1958) "Relational analysis: The study of social organizations
with survey methods", Human Organization 17:28--36.
}
\seealso{
Other segregation.measures: \code{\link{assort}},
  \code{\link{assort.default}},
  \code{\link{assort.igraph}}, \code{\link{assort.table}};
  \code{\link{ei}}, \code{\link{ei.array}},
  \code{\link{ei.igraph}}; \code{\link{freeman}},
  \code{\link{freeman.array}},
  \code{\link{freeman.igraph}}; \code{\link{gamix}},
  \code{\link{gamix.array}}, \code{\link{gamix.igraph}};
  \code{\link{orwg}}, \code{\link{orwg.array}},
  \code{\link{orwg.igraph}}; \code{\link{smi}},
  \code{\link{smi.array}}, \code{\link{smi.igraph}};
  \code{\link{ssi}}
}

