% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/EFnet.R
\name{EFnet}
\alias{EFnet}
\title{Example data from Echenique and Fryer (2006)}
\format{Object of class "igraph". An undirected network with vertex
attributes
\itemize{
\item \code{type} (values 1 or 2).
\item \code{name}
}}
\source{
Echenique, Federico and Roland G. Fryer, Jr. (2006) "A Measure of Segregation
Based On Social Interactions" Quarterly Journal of Economics CXXII(2):441-485
}
\description{
Artificial example data from Echenique & Fryer (2006) representing a city
with black and white neighbourhoods.
}
\details{
This data is taken from Echenique & Fryer (2006, figure III). The data
represent a fictional city composed of 30 neighborhoods that are either
black or white.
}
\examples{

set.seed(2992)
plot(EFnet, layout=layout.fruchterman.reingold,
    vertex.color=V(EFnet)$type+1, vertex.label.family="",
    sub="Source: Echenique & Fryer (2006)",
    main="Neighborhood racial segregation\\n in a fictional city" )
}
\keyword{datasets}

