% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ei}
\alias{ei}
\alias{ei.default}
\alias{ei.igraph}
\alias{ei.table}
\title{Krackhard and Stern's E-I index}
\usage{
ei(object, ...)

\method{ei}{table}(object, ...)

\method{ei}{igraph}(object, vattr, directed = is.directed(object),
  loops = any(is.loop(object)), ...)

\method{ei}{default}(object, ...)
}
\arguments{
\item{object}{R object, see Details for available methods}

\item{...}{other arguments passed to/from other methods}

\item{vattr}{character scalar or vector of length equal to the size of
\code{object}, vertex attribute for which mixing matrix is to be computed}

\item{directed}{logical whether the network is directed}

\item{loops}{logical, whether loops are allowed}
}
\value{
Numerical value of the E-I index.
}
\description{
An index proposed by Krackhard and Stern (1988) to capture relative
prevalence of between- and within-group ties. From that perspective it can
be interpreted as a measure of network segregation.
}
\details{
If \code{object} is a table, it is assumed to be a mixing matrix.

Method for igraphs

Default method tries to coerce \code{object} to table.
}
\examples{
ei(Wnet, "gender")
}
\seealso{
Other segregation.measures: \code{\link{assort}},
  \code{\link{assort.default}},
  \code{\link{assort.igraph}}, \code{\link{assort.table}};
  \code{\link{coleman}}, \code{\link{coleman.default}},
  \code{\link{coleman.igraph}},
  \code{\link{coleman.table}}; \code{\link{freeman}},
  \code{\link{freeman.default}},
  \code{\link{freeman.igraph}},
  \code{\link{freeman.table}}; \code{\link{gamix}},
  \code{\link{gamix.default}}, \code{\link{gamix.igraph}},
  \code{\link{gamix.table}}; \code{\link{orwg}},
  \code{\link{orwg.default}}, \code{\link{orwg.igraph}},
  \code{\link{orwg.table}}; \code{\link{smi}},
  \code{\link{smi.array}}, \code{\link{smi.igraph}};
  \code{\link{ssi}}
}

