% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ei}
\alias{ei}
\alias{ei.array}
\alias{ei.igraph}
\title{Krackhard and Stern's E-I index}
\usage{
ei(object, ...)

\method{ei}{array}(object, ...)

\method{ei}{igraph}(object, vattr, directed = is.directed(object),
  loops = any(is.loop(object)), ...)
}
\arguments{
\item{object}{R object, see Details for available methods}

\item{...}{other arguments passed to/from other methods}

\item{vattr}{character scalar or vector of length equal to the size of
\code{object}, vertex attribute for which mixing matrix is to be computed}
}
\value{
Numerical value of the E-I index.
}
\description{
An index proposed by Krackhard and Stern (1988) to capture relative
prevalence of between- and within-group ties. From that perspective it can
be interpreted as a measure of network segregation.
}
\details{
Method for mixing matrices

Method for igraphs
}
\examples{
ei(Wnet, "gender")
}
\seealso{
Other segregation.measures: \code{\link{assort}},
  \code{\link{assort.default}},
  \code{\link{assort.igraph}}, \code{\link{assort.matrix}};
  \code{\link{coleman}}, \code{\link{coleman.array}},
  \code{\link{coleman.igraph}}; \code{\link{freeman}},
  \code{\link{freeman.array}},
  \code{\link{freeman.igraph}}; \code{\link{gamix}},
  \code{\link{gamix.array}}, \code{\link{gamix.igraph}};
  \code{\link{orwg}}, \code{\link{orwg.array}},
  \code{\link{orwg.igraph}}; \code{\link{smi}},
  \code{\link{smi.array}}, \code{\link{smi.igraph}};
  \code{\link{ssi}}
}

