% Generated by roxygen2 (4.0.1): do not edit by hand
\name{freeman}
\alias{freeman}
\alias{freeman.igraph}
\alias{freeman.mixingm}
\title{Generalized Freeman's segregation index}
\usage{
freeman(object, ...)

\method{freeman}{mixingm}(object, dis = attr(object, "gsizes"),
  more = FALSE, ...)

\method{freeman}{igraph}(object, vattr, dis = NULL, ...)
}
\arguments{
\item{object}{R object, see Details for available methods}

\item{...}{other arguments passed to/from other methods}

\item{dis}{numeric, optional true distribution of types, see Details}

\item{more}{logical, should some more output be returned}

\item{vattr}{character scalar or any vector of length equal to
\code{vcount(object)}, name of the vertex attribute in \code{object}
designating the groups or a vector with the attribute itself}
}
\value{
The value of the Freeman's index.

If \code{more} is \code{TRUE}, some intermediate results are returned in a
list.
}
\description{
Calculate Freeman's segregation index for undirected netoworks with
arbitrary number of groups.
}
\details{
Freeman's segregation index (Freeman, 1978) is designed to capture the
extent to which the defined groups of vertices tend to have more edges with
vertices from the same group than with other groups.  Formally, the index
compares the observed number of between-group ties with the number of
between-group ties that would be expected if ties would be created randomly.

The index has some discontinuity as there are network and group configurations
that are characterized by the higher number of between-group ties that is
expected under a random graph. The index is truncates these situations and
takes a value of 0.

The original Freeman's formulation was for only two types of vertices.  Here
it is extended to the arbitrary number of types. The modification only
affects the way in which the expected number of inter-type edges under pure
random graph is calculated.

The function internally calculates the frequency of types of vertices in the
supplied attributer \code{vattr}. However, it is possible to override this
by specifying ``true'' type distribution with the \code{dis} argument. It is
assumed to be a table (as returned by \code{table}) or a numeric vector with
frequencies of types of vertices. This may be especially usefull when
dealing with large graphs with larger number of isolates.

Method for mixing matrices

Method for "igraph"s
}
\examples{
## White's data from Freeman's article
# segregation level
freeman(Wnet, "gender")
# using 'more' argument
freeman(Wnet, "gender", more=TRUE)
}
\references{
Freeman, Linton C. (1978) Segregation in Social Networks, \emph{Sociological
Methods & Research} \bold{6}(4):411--429
}
\seealso{
Other segregation.measures: \code{\link{assort}},
  \code{\link{assort.default}},
  \code{\link{assort.igraph}},
  \code{\link{assort.mixingm}}; \code{\link{coleman}},
  \code{\link{coleman.igraph}},
  \code{\link{coleman.mixingm}}; \code{\link{ei}},
  \code{\link{ei.igraph}}, \code{\link{ei.mixingm}};
  \code{\link{gamix}}, \code{\link{gamix.igraph}},
  \code{\link{gamix.mixingm}}; \code{\link{orwg}},
  \code{\link{orwg.igraph}}, \code{\link{orwg.mixingm}};
  \code{\link{smi}}, \code{\link{smi.igraph}},
  \code{\link{smi.mixingm}}; \code{\link{ssi}}
}
\keyword{graphs}

