% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{gender_roles}
\alias{gender_roles}
\title{Subset of Polish General Social Survey data with response to questions on
gender roles}
\format{Plain text file with header, space-sparated values, and the following columns}
\source{
Polish Social Science Data Archive (ADS)
}
\description{
Subset of Polish General Social Survey data with response to questions on
gender roles
}
\references{
PGSS reference
}
\keyword{datasets}

