% Generated by roxygen2 (4.0.1): do not edit by hand
\name{orwg}
\alias{orwg}
\alias{orwg.array}
\alias{orwg.igraph}
\title{Odds ratio of existence of within-group ties}
\usage{
orwg(object, ...)

\method{orwg}{array}(object, ...)

\method{orwg}{igraph}(object, vattr, ...)
}
\arguments{
\item{object}{R object, see Details for available methods}

\item{...}{other arguments passed to/from other methods}

\item{vattr}{character scalar or any vector, name of the vertex attribute or
the attribute itself (as a vector)}
}
\value{
Numeric value of the measure.
}
\description{
Odds ratio for connected, as opposed to disconnected, dyads depending
whether it is between- or within-group, i.e. how much more likely the dyad
will be connected if it is within-group.
}
\details{
The measure takes values, like all odds ratios, from (0; Inf).

Method for mixing matrices.

Method for igraphs
}
\examples{
orwg(Wnet, "gender")
}
\references{
Moody, Jim (2001) "Race, school integration, and friendship segregation in
America", American Journal of Sociology, 107(3):679--377
}
\seealso{
Other segregation.measures: \code{\link{assort}},
  \code{\link{assort.default}},
  \code{\link{assort.igraph}}, \code{\link{assort.matrix}};
  \code{\link{coleman}}, \code{\link{coleman.array}},
  \code{\link{coleman.igraph}}; \code{\link{ei}},
  \code{\link{ei.array}}, \code{\link{ei.igraph}};
  \code{\link{freeman}}, \code{\link{freeman.array}},
  \code{\link{freeman.igraph}}; \code{\link{gamix}},
  \code{\link{gamix.array}}, \code{\link{gamix.igraph}};
  \code{\link{smi}}, \code{\link{smi.array}},
  \code{\link{smi.igraph}}; \code{\link{ssi}}
}

