% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_redundancy.R
\name{pr_invdistance}
\alias{pr_invdistance}
\alias{pr_sharedp}
\alias{pairwise_redundancy}
\title{Pairwise tie redundancy}
\usage{
pr_invdistance(g, v = igraph::V(g))

pr_sharedp(g, v = igraph::V(g))

pairwise_redundancy(g, v = igraph::V(g))
}
\arguments{
\item{g}{igraph with vertex names}

\item{v}{vertex sequence of nodes to analyzed (as egos)}
}
\value{
\code{\link[=pr_invdistance]{pr_invdistance()}} returns a data frame with columns:
\itemize{
\item \code{ego} -- name of ego vertex
\item \code{v1} -- name of alter 1
\item \code{v2} -- name of alter 2
\item \code{invdis} -- inverse geodesic distance between nodes \code{v1} and \code{v2} in \code{g} having all ties of \code{ego} removed
}

Every function returns a data frame with columns \code{ego}, \code{v1},
\code{v2} with codes for, respectively, ego, alter1, and alter2. Additional column
depends on function and it is:
\describe{
\item{invdis}{inverse distance returned by \code{pr_invdistance}}
\item{sharedp}{number of shared partners returned by \code{pr_sharedp}}
\item{pr_redundancy}{the two above combined returned by \code{pairwise_redundancy}}
}
}
\description{
Measuring relative redundancy of ties in neighborhoods.

Function \code{\link[=pr_invdistance]{pr_invdistance()}} calculates for each ego in \code{v} an
inverse geodesic distance between alters \code{v1} and \code{v2} having all ego-alter
ties removed.

Function \code{\link[=pr_sharedp]{pr_sharedp()}} calculates the number of shared partners of
nodes \code{v1} and \code{v2} having all ties of \code{ego} removed.
\code{pr_sharedp} calculates number of shared partners of \code{v1} and
\code{v2} when all ego-alter ties are removed.

Function \code{\link[=pairwise_redundancy]{pairwise_redundancy()}} computes a combination of the to measures.
}
